Changed json response for backend yt and included all metrics to be real time
