+++ lampsz/apis/urls.py
@@ -34,7 +34,6 @@ urlpatterns += [
     path("profile/<int:user_id>", profile.public_user_detail),
     path("influencer/<int:user_id>", profile.influencer_edit_view),
     path("company/<int:user_id>", profile.company_edit_view),
-    path("create_task/", profile.create_marketing_task),
 ]
 
 urlpatterns += [
