+++ lampsz/apis/views/profile.py
@@ -1,9 +1,9 @@
-from django.http import JsonResponse
 from rest_framework import status
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
+from rest_framework.response import Response
 
-from lampsz.apis import models, serializers, utils
+from lampsz.apis import models, serializers, services, utils
 
 
 @api_view(["GET"])
@@ -16,7 +16,7 @@ def public_user_detail(request, user_id):
         else:
             company = models.Company.objects.filter(user=user).first()
     except models.Influencer.DoesNotExist or models.Company.DoesNotExist or models.User.DoesNotExist:
-        return JsonResponse(
+        return Response(
             {"message": "The User does not exist"},
             status=status.HTTP_404_NOT_FOUND,
         )
@@ -24,13 +24,13 @@ def public_user_detail(request, user_id):
         influencer_serializer = serializers.InfluencerSerializer(influencer, many=False)
         data = dict(influencer_serializer.data)
         data["userType"] = user.get_user_type().value
-        utils.get_youtube_channel_detail(data, influencer.channel_id)
-        return JsonResponse(data, status=status.HTTP_200_OK)
+        services.get_youtube_channel_detail(data, influencer.channel_id)
+        return Response(data, status=status.HTTP_200_OK)
     else:
         company_serializer = serializers.CompanySerializer(company, many=False)
         data = dict(company_serializer.data)
         data["userType"] = user.get_user_type().value
-        return JsonResponse(data, status=status.HTTP_200_OK)
+        return Response(data, status=status.HTTP_200_OK)
 
 
 @api_view(["PUT"])
@@ -40,19 +40,19 @@ def influencer_edit_view(request, user_id):
         user = models.User.objects.get(id=user_id)
         influencer = models.Influencer.objects.filter(user=user).first()
     except models.Influencer.DoesNotExist:
-        return JsonResponse(
+        return Response(
             {"message": "The Influencer does not exist"},
             status=status.HTTP_404_NOT_FOUND,
         )
     if influencer.user.id != request.user.id:
-        return JsonResponse(
+        return Response(
             {"message": "This user is not authorized to access this data"},
             status=status.HTTP_401_UNAUTHORIZED,
         )
     influencer_data = request.data
     influencer_serializer = serializers.InfluencerSerializer(influencer, many=False)
     influencer_serializer.update(influencer, influencer_data)
-    return JsonResponse(influencer_serializer.data, status=status.HTTP_200_OK)
+    return Response(influencer_serializer.data, status=status.HTTP_200_OK)
 
 
 @api_view(["PUT"])
@@ -62,40 +62,15 @@ def company_edit_view(request, user_id):
         user = models.User.objects.get(id=user_id)
         company = models.Company.objects.filter(user=user).first()
     except models.Company.DoesNotExist:
-        return JsonResponse(
+        return Response(
             {"message": "The Company does not exist"}, status=status.HTTP_404_NOT_FOUND
         )
     if company.user.id != request.user.id:
-        return JsonResponse(
+        return Response(
             {"message": "This user is not authorized to access this data"},
             status=status.HTTP_401_UNAUTHORIZED,
         )
     company_data = request.data
     company_serializer = serializers.CompanySerializer(company, many=False)
     company_serializer.update(company, company_data)
-    return JsonResponse(company_serializer.data, status=status.HTTP_200_OK)
-
-
-@api_view(["POST"])
-def create_marketing_task(request):
-    userId = request.data["userId"]
-    data = {}
-    data["company"] = models.Company.objects.filter(user_id=userId)[0].id
-
-    data["title"] = request.data["title"]
-    data["description"] = request.data["description"]
-    data["deliverables"] = request.data["deliverables"]
-    data["compensation"] = float(request.data["compensation"])
-    data["posted_date"] = request.data["postedDate"]
-    data["end_date"] = request.data["endDate"]
-    data["location"] = request.data["location"]
-    data["image"] = request.data["image"]
-
-    serializer = serializers.MarketingTaskSerializer(data=data)
-
-    if serializer.is_valid():
-        serializer.save()
-    else:
-        print(serializer.errors)
-
-    return Response({"message": "successful"}, status=200)
+    return Response(company_serializer.data, status=status.HTTP_200_OK)
